requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.14,<4.0
